Updated CelestialBody class with 3D pos and vel
Added the 3rd coordinate and a new member "velocity" to CelestialBody
class, updating getters and setters accordingly; also added methods
updatePos and updateVel, that will be crucial to the simulation.
Fixed the CelestialBody constructor which did not set m_name.

Things to do: consider if adding children classes Star, Planet and Moon
is useful or just bloat, as they wouldn't have any peculiarity compared
to CelestialBody.